+++ db/db_test.cc
@@ -4689,7 +4689,7 @@ TEST_F(DBTest, DynamicUniversalCompactionOptions) {
   DestroyAndReopen(options);
 
   // Initial defaults
-  ASSERT_EQ(dbfull()->GetOptions().compaction_options_universal.size_ratio, static_cast<uint>(1));
+  ASSERT_EQ(dbfull()->GetOptions().compaction_options_universal.size_ratio, 1U);
   ASSERT_EQ(dbfull()->GetOptions().compaction_options_universal.min_merge_width,
             2);
   ASSERT_EQ(dbfull()->GetOptions().compaction_options_universal.max_merge_width,
